Change `Value::Vector` to hold `Vec<Value>`.
This is a breaking version change, because the underlying type of the field of `Value::Vector`
was (and still is) exposed and is changed.

This is arguably more useful than it holding `Box<[Value]>`, because a user can now:

- mutate the underlying `Vec` as desired, without incurring extra costs of converting between
  boxed slice (converting `Vec<T>` to `Box<[T]>` can incur an extra allocation and copy, to shrink
  capacity to be equal to len).

- supply a `Vec` with desired `capacity > len` and have this capacity kept (e.g. for later
  mutations of it).

Also, this enables a more-efficient implementation of `DeepSafeDrop::take_next_child_at_pos_index`
for `Value::Vector`.